so it turns out that implementing a seed fill algorithm while making no
so it turns out that implementing a seed fill algorithm while making
no assumptions about memory layout or data precision is not exactly
trivial.  this is a first cut, with magic wand and bucket applications.

ray lehtiniemi <rayl@netrover.com>

+++ app/gimprc.h
@@ -44,7 +44,7 @@ extern int       levels_of_undo;
 extern int       color_cube_shades[];
 extern int       install_cmap;
 extern int       cycled_marching_ants;
-extern int       default_threshold;
+extern double    default_threshold;
 extern int       stingy_memory_use;
 extern int       allow_resize_windows;
 extern int       no_cursor_updating;
